+++ modules/core/migrations/m240822_010917_data_his.php
@@ -18,7 +18,7 @@ class m240822_010917_data_his extends Migration
             'data'       => $this->text()->comment('数据'),
             'color'      => $this->string()->comment('颜色'),
             'user_id'    => $this->bigInteger()->comment('用户ID'),
-            'created_at' => $this->integer()->comment('创建时间'),
+            'created_at' => $this->bigInteger()->comment('创建时间'),
         ]);
     }
