+++ modules/core/migrations/m240822_010917_mail.php
@@ -17,7 +17,7 @@ class m240822_010917_mail extends Migration
             'name'    => $this->string()->comment('邮件模板名称'),
             'key'     => $this->string()->comment('邮件模板键名')->defaultValue(''),
             'content' => $this->text()->comment('邮件模板内容'), 
-            'created_at' => $this->integer()->comment('创建时间'),
+            'created_at' => $this->bigInteger()->comment('创建时间'),
         ]);
     }
